--HG-- rename : UModSys/src/module/2d/lib2d/umodsys.mod.2d.lib2d.convert.h => UModSys/src/module/2d/stdlib/umodsys.mod.2d.stdlib.convert.h rename : UModSys/src/module/2d/lib2d/umodsys.mod.2d.lib2d.convertfast.h => UModSys/src/module/2d/stdlib/umodsys.mod.2d.stdlib.convertfast.h rename : UModSys/src/module/2d/lib2d/umodsys.mod.2d.lib2d.convertslow.h => UModSys/src/module/2d/stdlib/umodsys.mod.2d.stdlib.convertslow.h rename : UModSys/src/module/2d/lib2d/umodsys.mod.2d.lib2d.cpp => UModSys/src/module/2d/stdlib/umodsys.mod.2d.stdlib.cpp rename : UModSys/src/module/2d/lib2d/umodsys.mod.2d.lib2d.h => UModSys/src/module/2d/stdlib/umodsys.mod.2d.stdlib.h rename : UModSys/src/module/2d/lib2d/umodsys.mod.2d.lib2d.memoryimage.h => UModSys/src/module/2d/stdlib/umodsys.mod.2d.stdlib.memoryimage.h rename : UModSys/src/module/2d/lib2d/version_module.h => UModSys/src/module/2d/stdlib/version_module.h extra : convert_revision : svn%3Ac7b0911a-351f-11e0-b982-4926ce03a894%40538

+++ src/module/2d/stdlib/umodsys.mod.2d.stdlib.memoryimage.h
@@ -42,7 +42,7 @@ public:
         s.offsets[s.nplanes] = binsize;
         s.infos[s.nplanes].set(eImageType(inf.type & it_Mask_Plane1), inf.size);
         //
-        xs = (1<<(UMODSYS_ARCH_BYTE_BITS * ps));
+        xs = size_t(1<<(UMODSYS_ARCH_BYTE_BITS * int(ps)));
         es = GetPixelSize(inf.type, 1) * size_t(xs);
         binsize += es;
         s.nplanes++;
